refactor: Replace deprecated navigation and presentation logic with new domain-driven implementation
*   Deletes `ke.don.ski.navigation.deprecated` and `ke.don.ski.presentation.deprecated` packages.
*   Moves `DeckMode` enum to `ke.don.ski.domain` and introduces `LocalDeckMode` as a `CompositionLocal`.
*   Relocates `ToolBar` and `TableOfContent` components to `ke.don.ski.presentation.ui`.
*   Updates `DeckScaffolding` to consume `DeckMode` via `LocalDeckMode.current` and use the updated `DeckNavigator`.
*   Enhances `SlideConfig.Render` to apply consistent padding and conditionally show the timer based on `DeckMode`.
*   Updates `IntroductionScreen` layout to use `Arrangement.SpaceBetween` and removes explicit padding.
*   Modifies the `Deck` entry point to use `CompositionLocalProvider` for theme and mode state.
*   Fixes window close behavior in `jvmMain` to properly handle application exit for single-window launches.
*   Sets a default `Fade` transition for the "Introduction" slide in `skiPresentationSlides`.
