+++ composeApp/src/commonMain/kotlin/ke/don/ski/domain/DeckMode.kt
@@ -1,0 +1,3 @@
+package ke.don.ski.domain
+
+enum class DeckMode { Presenter, Local }
