+++ composeApp/src/commonMain/kotlin/ke/don/ski/domain/LocalDeckMode.kt
@@ -1,0 +1,7 @@
+package ke.don.ski.domain
+
+import androidx.compose.runtime.staticCompositionLocalOf
+
+val LocalDeckMode = staticCompositionLocalOf<DeckMode> {
+    error("DeckMode not provided")
+}
